services/auth-service/requirements.txt
Django==5.1.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.2
django-allauth==65.0.0
dj-rest-auth[with_social]==7.0.1
python-social-auth==0.3.6
social-auth-app-django==5.4.1
django-environ==0.11.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
uvicorn==0.30.1